requirements.txt
PyQt6==6.6.1
psutil==5.9.6
pywin32==306
Pillow>=10.0.0

